terminado y testeado

+++ src/main/java/ar/edu/unahur/obj2/command/tareas/Info.java
@@ -5,16 +5,16 @@ import ar.edu.unahur.obj2.command.robot.Robot;
 public class Info extends Comando {
 
     @Override
-    public Integer consumo() {
-        return 0;
+    public Double consumo() {
+        return 0.0;
  }
     @Override
     public void doEjecutar(Robot robot) {
         System.out.println(robot.promedioTiempo());
  }
     @Override
-    public Integer duracion() {
-       return 0;
+    public Double duracion() {
+       return 0.0;
     }
 
 }
