+++ src/main/java/ar/edu/unahur/obj2/command/tareas/Tarea.java
@@ -4,8 +4,8 @@ package ar.edu.unahur.obj2.command.tareas;
 import ar.edu.unahur.obj2.command.robot.Robot;
 
 public interface Tarea {
-    public Integer consumo();
-    public Integer duracion();
+    public Double consumo();
+    public Double duracion();
     public void ejecutar(Robot robot);
     public void doEjecutar(Robot robot);
 }
